+++ modules/kafka-distributed-processing/src/main/scala/aecor/kafkadistributedprocessing/DistributedProcessing.scala
@@ -2,7 +2,8 @@ package aecor.kafkadistributedprocessing
 
 import java.util.Properties
 
-import aecor.kafkadistributedprocessing.Kafka._
+import aecor.kafkadistributedprocessing.internal.Kafka
+import aecor.kafkadistributedprocessing.internal.Kafka._
 import cats.effect.{ ConcurrentEffect, ContextShift, Timer }
 import cats.implicits._
 import cats.effect.implicits._
@@ -11,22 +12,22 @@ import org.apache.kafka.clients.consumer.ConsumerConfig
 
 final class DistributedProcessing(settings: DistributedProcessingSettings) {
 
-  private def assignRange(size: Int, partitionCount: Int, partition: Int): (Int, Int) = {
+  private def assignRange(size: Int, partitionCount: Int, partition: Int): Option[(Int, Int)] = {
     val even = size / partitionCount
     val reminder = size % partitionCount
     if (partition >= partitionCount) {
-      (0, 0)
+      none
     } else {
       if (partition < reminder) {
-        (partition * (even + 1), even + 1)
-      } else {
-        (reminder + partition * even, even)
-      }
+        (partition * (even + 1), even + 1).some
+      } else if (even > 0) {
+        (reminder + partition * even, even).some
+      } else none
     }
   }
 
   /**
-    * Starts `processes` distributed over internal Kafka topic consumers.
+    * Starts `processes` distributed over internal Kafka topic consumer.
     *
     * @param name - used as groupId for underlying Kafka partition assignment machinery
     * @param processes - list of processes to distribute
@@ -38,20 +39,18 @@ final class DistributedProcessing(settings: DistributedProcessingSettings) {
       .assignPartitions(settings.asProperties(name), settings.topicName)
       .parEvalMapUnordered(Int.MaxValue) {
         case AssignedPartition(partition, partitionCount, watchRevocation, release) =>
-          val (offset, processCount) = assignRange(processes.size, partitionCount, partition)
-
-          val task =
-            if (processCount > 0)
+          assignRange(processes.size, partitionCount, partition).fold(release) {
+            case (offset, processCount) =>
               Stream
                 .range[F](offset, offset + processCount)
                 .parEvalMapUnordered(processCount)(processes)
                 .compile
                 .drain
-            else
-              ().pure[F]
-          task.race(watchRevocation).flatMap {
-            case Left(_)      => release
-            case Right(value) => value
+                .race(watchRevocation)
+                .flatMap {
+                  case Left(_)         => release
+                  case Right(callback) => callback
+                }
           }
       }
       .compile
