+++ modules/kafka-distributed-processing/src/main/scala/aecor/kafkadistributedprocessing/internal/RebalanceEvents.scala
@@ -1,4 +1,4 @@
-package aecor.kafkadistributedprocessing
+package aecor.kafkadistributedprocessing.internal
 
 import java.util
